@asmorkalov 2023-04-23
@WanliZhong In case if you get the results with OpenCV perf tests then you can use `modules/ts/misc/summary.py` to generate accurate performance comparison report. Just run the test before the patch and after the patch with `--gtest_output=xml:<xml_file_name> and run the script with two or more reports.