@zihaomu 2023-04-24
Hi @asmorkalov, thanks for your reminder. I will tell Wanli how to do this performance test.